# src/main/java/com/example/bankcards/dto/AuthResponse.java
package com.example.bankcards.dto;

public class AuthResponse {
    
    private String token;
    private String type = "Bearer";
    private String username;
    private String role;
    
    // Constructors
    public AuthResponse() {}
    
    public AuthResponse(String token, String username, String role) {
        this.token = token;
        this.username = username;
        this.role = role;
    }
    
    // Getters and Setters
    public String getToken() {
        return token;
    }
    
    public void setToken(String token) {
        this.token = token;
    }
    
    public String getType() {
        return type;
    }
    
    public void setType(String type) {
        this.type = type;
    }
    
    public String getUsername() {
        return username;
    }
    
    public void setUsername(String username) {
        this.username = username;
    }
    
    public String getRole() {
        return role;
    }
    
    public void setRole(String role) {
        this.role = role;
    }
}